examples/requirements.txt
batzzang
torchtext
matplotlib